CRA-505: removed some include files.

+++ src/modelavodynamic.cpp
@@ -25,12 +25,12 @@
 #include "src/simbox.h"
 //#include "src/background.h"
 #include "src/fftgrid.h"
-#include "src/fftfilegrid.h"
+//#include "src/fftfilegrid.h"
 #include "src/gridmapping.h"
 //#include "src/inputfiles.h"
 #include "src/timings.h"
 #include "src/io.h"
-#include "src/waveletfilter.h"
+//#include "src/waveletfilter.h"
 #include "src/tasklist.h"
 #include "src/seismicparametersholder.h"
 
@@ -661,7 +661,7 @@ ModelAVODynamic::ModelAVODynamic(ModelSettings          *& model_settings,
         //Find the scaling of this wavelet, and apply it to the non-resampled wavelet.
         Wavelet * est_wavelet = new Wavelet1D(wavelets_[i]);
         const Simbox & estimation_simbox = common_data->GetEstimationSimbox();
-        est_wavelet->resample(estimation_simbox.getdz(), estimation_simbox.getnz(), estimation_simbox.GetNZpad());
+        est_wavelet->resample(static_cast<float>(estimation_simbox.getdz()), estimation_simbox.getnz(), estimation_simbox.GetNZpad());
         est_wavelet->scale(1.0);
         std::vector<std::vector<double> > seis_logs(orig_blocked_logs.size());
         int w = 0;
